# src/main/java/pavel/lab/cinema/dto/error/ErrorResponse.java
package pavel.lab.cinema.dto.error;

import com.fasterxml.jackson.annotation.JsonInclude;

import java.time.LocalDateTime;

@JsonInclude(JsonInclude.Include.NON_NULL)
public record ErrorResponse(
        int status,
        String message,
        LocalDateTime time
) {}