# src/main/java/pavel/lab/cinema/dto/requestdto/GenreRequestDTO.java
package pavel.lab.cinema.dto.requestdto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class GenreRequestDTO {

    @NotBlank
    @Size(min = 1, max = 20)
    private String name;
}