# src/main/java/pavel/lab/cinema/dto/requestdto/MovieRequestDTO.java
package pavel.lab.cinema.dto.requestdto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class MovieRequestDTO {

    @NotBlank
    @Size(min = 1, max = 50)
    private String title;

    @NotNull
    @Positive
    private Integer duration;

    @NotNull
    private List<Long> genreIds;
}