# src/main/resources/static/api-docs.yaml
openapi: 3.1.0
info:
  title: Cinema API
  description: REST API для управления кинотеатром — фильмы, сеансы, залы, билеты, жанры и посетители.
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
paths:
...
  /genres:
    get:
      tags: [genres]
      summary: Получить все жанры
      description: Возвращает список всех жанров без пагинации.
      operationId: getAllGenres
      responses:
        "200":
          description: Список жанров успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GenreDTO"
              example:
                - id: 1
                  name: Боевик
                - id: 2
                  name: Комедия
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /genres/{id}:
    get:
      tags: [genres]
      summary: Получить жанр по ID
      description: Возвращает жанр по указанному идентификатору. Если жанр не найден — возвращает 404.
      operationId: getGenreById
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Жанр найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreDTO"
              example:
                id: 1
                name: Боевик
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /genres/post:
    post:
      tags: [genres]
      summary: Создать жанр
      description: Создаёт новый жанр. Название должно быть от 1 до 20 символов.
      operationId: createGenre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenreRequestDTO"
            example:
              name: Триллер
      responses:
        "200":
          description: Жанр успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreDTO"
              example:
                id: 3
                name: Триллер
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/DataIntegrityError"
  /genres/update/{id}:
    put:
      tags: [genres]
      summary: Обновить жанр
      description: Обновляет название жанра по ID. Название должно быть от 1 до 20 символов.
      operationId: updateGenre
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenreRequestDTO"
            example:
              name: Мелодрама
      responses:
        "200":
          description: Жанр успешно обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreDTO"
              example:
                id: 1
                name: Мелодрама
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /genres/delete/{id}:
    delete:
      tags: [genres]
      summary: Удалить жанр
      description: Удаляет жанр по ID. Жанр автоматически отвязывается от всех фильмов перед удалением.
      operationId: deleteGenre
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Жанр успешно удалён
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─── HALLS ─────────────────────────────────────────────────────────────────
  /halls:
    get:
      tags: [halls]
      summary: Получить все залы
      description: Возвращает список всех кинозалов без пагинации.
      operationId: getAllHalls
      responses:
        "200":
          description: Список залов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HallDTO"
              example:
                - id: 1
                  name: Красный зал
                  price: 350.0
                  seatAmount: 100
                - id: 2
                  name: Синий зал
                  price: 500.0
                  seatAmount: 60
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /halls/{id}:
    get:
      tags: [halls]
      summary: Получить зал по ID
      description: Возвращает кинозал по указанному идентификатору.
      operationId: getHallById
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Зал найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HallDTO"
              example:
                id: 1
                name: Красный зал
                price: 350.0
                seatAmount: 100
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /halls/post:
    post:
      tags: [halls]
      summary: Создать зал
      description: Создаёт новый кинозал. Цена и количество мест должны быть положительными числами.
      operationId: createHall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HallRequestDTO"
            example:
              name: VIP зал
              price: 800.0
              seatAmount: 30
      responses:
        "200":
          description: Зал успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HallDTO"
              example:
                id: 3
                name: VIP зал
                price: 800.0
                seatAmount: 30
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/DataIntegrityError"
  /halls/update/{id}:
    put:
      tags: [halls]
      summary: Обновить зал
      description: Обновляет данные кинозала по ID.
      operationId: updateHall
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HallRequestDTO"
            example:
              name: VIP зал
              price: 900.0
              seatAmount: 25
      responses:
        "200":
          description: Зал успешно обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HallDTO"
              example:
                id: 3
                name: VIP зал
                price: 900.0
                seatAmount: 25
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /halls/delete/{id}:
    delete:
      tags: [halls]
      summary: Удалить зал
      description: Удаляет кинозал по ID.
      operationId: deleteHall
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Зал успешно удалён
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─── MOVIES ────────────────────────────────────────────────────────────────
  /movies:
    get:
      tags: [movies]
      summary: Получить все фильмы (с пагинацией)
      description: Возвращает постраничный список фильмов. По умолчанию размер страницы — 3.
      operationId: getAllMovies
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Список фильмов успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseMovieDTO"
              example:
                content:
                  - id: 1
                    title: Интерстеллар
                    duration: 169
                    genres: [Фантастика, Драма]
                  - id: 2
                    title: Джокер
                    duration: 122
                    genres: [Триллер, Драма]
                metadata:
                  page: 0
                  totalPage: 5
                  size: 3
                  totalElements: 14
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /movies/{id}:
    get:
      tags: [movies]
      summary: Получить фильм по ID
      description: Возвращает фильм по указанному идентификатору.
      operationId: getMovieById
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Фильм найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDTO"
              example:
                id: 1
                title: Интерстеллар
                duration: 169
                genres: [Фантастика, Драма]
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /movies/post:
    post:
      tags: [movies]
      summary: Создать фильм
      description: >
        Создаёт новый фильм. Необходимо передать список ID существующих жанров.
        Если хотя бы один жанр не найден — возвращается 404.
      operationId: createMovie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieRequestDTO"
            example:
              title: Дюна
              duration: 155
              genreIds: [1, 3]
      responses:
        "200":
          description: Фильм успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDTO"
              example:
                id: 5
                title: Дюна
                duration: 155
                genres: [Фантастика, Боевик]
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /movies/update/{id}:
    put:
      tags: [movies]
      summary: Обновить фильм
      description: Обновляет данные фильма по ID, включая список жанров.
      operationId: updateMovie
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieRequestDTO"
            example:
              title: Дюна 2
              duration: 166
              genreIds: [1, 3]
      responses:
        "200":
          description: Фильм успешно обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDTO"
              example:
                id: 5
                title: Дюна 2
                duration: 166
                genres: [Фантастика, Боевик]
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /movies/delete/{id}:
    delete:
      tags: [movies]
      summary: Удалить фильм
      description: Удаляет фильм по ID. Очищает кэш фильмов после удаления.
      operationId: deleteMovie
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Фильм успешно удалён
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─── SESSIONS ──────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: Уникальный идентификатор ресурса
      schema:
        type: integer
        format: int64
        example: 1
    PageParam:
      name: page
      in: query
      required: false
      description: Номер страницы (начиная с 0)
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
        example: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Количество элементов на странице (по умолчанию 3)
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 3
        example: 3
    SortParam:
      name: sort
      in: query
      required: false
      description: "Сортировка в формате: поле,направление (например: id,asc)"
      schema:
        type: array
        items:
          type: string
        example: ["id,asc"]
  responses:
    BadRequest:
      description: >
        Некорректный запрос (400). Возможные причины:
        ошибка валидации полей (`@Valid`),
        неверный формат JSON,
        неправильный тип параметра,
        нарушение бизнес-логики (например, номер места превышает вместимость зала).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validationError:
              summary: Ошибка валидации входных данных
              value:
                status: 400
                message: "Некорректный запрос: ошибка валидации входных данных"
                time: "2025-06-15T12:00:00"
            typeMismatch:
              summary: Неправильный тип параметра
              value:
                status: 400
                message: "Некорректный запрос: неправильный тип передаваемого параметра"
                time: "2025-06-15T12:00:00"
            invalidJson:
              summary: Ошибка формата JSON
              value:
                status: 400
                message: "Некорректный запрос: ошибка в формате JSON-сообщения"
                time: "2025-06-15T12:00:00"
            businessLogic:
              summary: Нарушение бизнес-логики (IllegalArgumentException)
              value:
                status: 400
                message: "Некорректный запрос: в зале Красный зал недостаточно мест"
                time: "2025-06-15T12:00:00"
    NotFound:
      description: Ресурс не найден (404). Возникает при обращении к несуществующей сущности.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 404
            message: "Фильм с ID 99 не найден"
            time: "2025-06-15T12:00:00"
    MethodNotAllowed:
      description: >
        Неправильный HTTP-метод (405).
        Возникает при обращении к эндпоинту с недопустимым методом
        (например, POST вместо GET).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 405
            message: "Неправильный тип запроса"
            time: "2025-06-15T12:00:00"
    DataIntegrityError:
      description: >
        Нарушение целостности данных (500).
        Возникает при нарушении ограничений БД — уникальности, внешних ключей и т.п.
        (`DataIntegrityViolationException`, `ConstraintViolationException`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 500
            message: "Нарушение целостности входных данных"
            time: "2025-06-15T12:00:00"
    InternalServerError:
      description: >
        Непредвиденная ошибка сервера (500).
        Возникает при любом необработанном исключении.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 500
            message: "Непредвиденная ошибка на сервере"
            time: "2025-06-15T12:00:00"
  schemas:
...
    GenreRequestDTO:
      type: object
      description: Тело запроса для создания или обновления жанра
      required: [name]
      properties:
        name:
          type: string
          description: Название жанра (от 1 до 20 символов)
          minLength: 1
          maxLength: 20
          example: Фантастика
    HallRequestDTO:
      type: object
      description: Тело запроса для создания или обновления кинозала
      required: [name, price, seatAmount]
      properties:
        name:
          type: string
          description: Название зала (от 1 до 20 символов)
          minLength: 1
          maxLength: 20
          example: VIP зал
        price:
          type: number
          format: double
          description: Цена билета в зале (положительное число)
          exclusiveMinimum: 0
          example: 800.0
        seatAmount:
          type: integer
          format: int32
          description: Количество мест в зале (положительное целое число)
          exclusiveMinimum: 0
          example: 30
    MovieRequestDTO:
      type: object
      description: Тело запроса для создания или обновления фильма
      required: [title, duration, genreIds]
      properties:
        title:
          type: string
          description: Название фильма (от 1 до 50 символов)
          minLength: 1
          maxLength: 50
          example: Интерстеллар
        duration:
          type: integer
          format: int32
          description: Длительность фильма в минутах (положительное число)
          exclusiveMinimum: 0
          example: 169
        genreIds:
          type: array
          description: Список ID жанров, привязанных к фильму
          items:
            type: integer
            format: int64
          example: [1, 3]
...
    GenreDTO:
      type: object
      description: Данные жанра
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор жанра
          example: 1
        name:
          type: string
          description: Название жанра
          example: Фантастика
    HallDTO:
      type: object
      description: Данные кинозала
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор зала
          example: 1
        name:
          type: string
          description: Название зала
          example: Красный зал
        price:
          type: number
          format: double
          description: Цена билета в данном зале
          example: 350.0
        seatAmount:
          type: integer
          format: int32
          description: Количество мест в зале
          example: 100
    MovieDTO:
      type: object
      description: Данные фильма
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор фильма
          example: 1
        title:
          type: string
          description: Название фильма
          example: Интерстеллар
        duration:
          type: integer
          format: int32
          description: Длительность фильма в минутах
          example: 169
        genres:
          type: array
          description: Список названий жанров фильма
          items:
            type: string
          example: [Фантастика, Драма]
...
    Metadata:
      type: object
      description: Метаданные пагинации
      properties:
        page:
          type: integer
          format: int32
          description: Текущий номер страницы (начиная с 0)
          example: 0
        totalPage:
          type: integer
          format: int32
          description: Общее количество страниц
          example: 5
        size:
          type: integer
          format: int32
          description: Размер страницы
          example: 3
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов во всех страницах
          example: 14
    PageResponseMovieDTO:
      type: object
      description: Постраничный ответ со списком фильмов
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/MovieDTO"
        metadata:
          $ref: "#/components/schemas/Metadata"
...
    ErrorResponse:
      type: object
      description: Стандартный ответ при ошибке
      properties:
        status:
          type: integer
          format: int32
          description: HTTP-код ошибки
          example: 404
        message:
          type: string
          description: Человекочитаемое описание ошибки
          example: "Фильм с ID 5 не найден"
        time:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2025-06-15T12:00:00"